manage add and get services for car, rental and user

+++ RentalCarsWithRepositoryDesignPattern/Domain/Data/User.cs
@@ -5,7 +5,7 @@ public class User
     public long Id { get; set; }
     public string Name { get; set; }
     public string LastName { get; set; }
-    public long Age { get; set; }
+    public int Age { get; set; }
     public decimal MinimumBalance { get; set; }
     public string TaxIdNumber { get; set; }
 }
